+++ _test/example/append_body/index.html
@@ -1,0 +1,7 @@
+<html>
+
+<head>
+  <script defer src="main.dart.js"></script>
+</head>
+
+</html>
