--- webdev/test/serve/injected/fixtures/pubspec.yaml
@@ -1,9 +1,0 @@
-name: fixture
-description: Dummy app for tests.
-
-environment:
-  sdk: ">=2.0.0 <3.0.0"
-
-dev_dependencies:
-  build_runner: '>=0.10.0 <2.0.0'
-  build_web_compilers: ^2.0.0
