--- webdev/test/serve/injected/fixtures/web/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>Test App</title>
-</head>
-
-<body>
-    <script type="application/javascript" src="main.dart.js"></script>
-</body>
-
-</html>
